#LUC070-113 Adde retry functionality to deposit file size check and copy from user space.   Also added it to the retrieve free space check too
1) Added new BaseTwoSpeedRetryTask.java class to hold the TwoSpeedConfig methods used by both retrieve and deposit.
2) Updated the Deposit calculateTotalDepositSize & copyFromUserFS methods to use the twoSpeedRetryTemplate code
3) Updated the Retrieve calculateFreeSpace method to use the twoSpeedRetryTemplate code
4) Updated the RetrieveTest so the test for failing to calculate size now expects the correct error messsge.

I still need to add tests to the DepositTest (which I assume exists)
