+++ datavault-worker/src/main/java/org/datavaultplatform/worker/tasks/Deposit.java
@@ -9,19 +9,17 @@ import org.datavaultplatform.common.event.InitStates;
 import org.datavaultplatform.common.event.UpdateProgress;
 import org.datavaultplatform.common.event.deposit.*;
 import org.datavaultplatform.common.io.Progress;
-import org.datavaultplatform.common.storage.ArchiveStore;
-import org.datavaultplatform.common.storage.Device;
-import org.datavaultplatform.common.storage.UserStore;
-import org.datavaultplatform.common.storage.Verify;
+import org.datavaultplatform.common.storage.*;
 import org.datavaultplatform.common.task.Context;
-import org.datavaultplatform.common.task.Task;
 import org.datavaultplatform.common.task.TaskExecutor;
 import org.datavaultplatform.common.util.StorageClassNameResolver;
 import org.datavaultplatform.common.util.StorageClassUtils;
 import org.datavaultplatform.common.util.Utils;
 import org.datavaultplatform.worker.operations.*;
+import org.datavaultplatform.worker.retry.TwoSpeedRetry;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.retry.RetryCallback;
 
 import java.io.File;
 import java.nio.file.Path;
@@ -30,7 +28,7 @@ import java.util.*;
 /**
  * A class that extends Task which is used to handle Deposits to the vault
  */
-public class Deposit extends Task {
+public class Deposit extends BaseTwoSpeedRetryTask {
 
     private static final Logger logger = LoggerFactory.getLogger(Deposit.class);
     private static final Set<String> RESTART_FROM_BEGINNING = new HashSet<>();
@@ -124,6 +122,7 @@ public class Deposit extends Task {
         }
         
         this.initStates();
+        this.setupUserFsTwoSpeedRetry(properties);
         
         eventSender.send(new Start(jobID, depositId)
             .withUserId(userID)
@@ -224,7 +223,11 @@ public class Deposit extends Task {
             // Ask the driver to copy files to our working directory
             logger.debug("CopyFromUserStorage filePath:" + filePath);
             logger.debug("CopyFromUserStorage outputFile:" + outputFile.getAbsolutePath());
-            ((Device)userStore).retrieve(filePath, outputFile, progress);
+            TwoSpeedRetry.DvRetryTemplate template = userFsTwoSpeedRetry.getRetryTemplate(String.format("fromUserFs - %s", filePath));
+            template.execute((RetryCallback<Object, Exception>) retryContext -> {
+                ((Device)userStore).retrieve(filePath, outputFile, progress);
+                return null;
+            });
         } finally {
             // Stop the tracking thread
             tracker.stop();
@@ -474,19 +477,9 @@ public class Deposit extends Task {
     private void verifyTarFile(Path tempPath, File tarFile, String origTarHash) throws Exception {
 
         Utils.checkFileHash("tar", tarFile, origTarHash);
-        // Decompress to the temporary directory
-        //File bagDir = Tar.unTar(tarFile, tempPath);
-        
-        // Validate the bagit directory
-        //if (!Packager.validateBag(bagDir)) {
-        //    throw new Exception("Bag is invalid");
-        //} else {
-        //    logger.info("Bag is valid");
-        //}
         
         // Cleanup
         logger.info("Cleaning up ...");
-        //FileUtils.deleteDirectory(bagDir);
         tarFile.delete();
     }
     
@@ -521,14 +514,10 @@ public class Deposit extends Task {
 
                 try {
                     UserStore userStore = userStores.get(storageID);
-                    long fileSize = userStore.getSize(storagePath);
-                    //logger.info("FileSize = '" + fileSize + "'");
-                    //if (fileSize == 0) {
-                    //    String msg = "Deposit failed: file size is 0";
-                    //    logger.error(msg);
-                    //    eventSender.send(new Error(jobID, depositId, msg).withUserId(userID));
-                    //    throw new RuntimeException(new Exception(msg));
-                    //}
+                    TwoSpeedRetry.DvRetryTemplate template = userFsTwoSpeedRetry.getRetryTemplate(String.format("calcSizeFromFS - %s", filePath));
+                    long fileSize = (long)template.execute((RetryCallback<Object, Exception>) retryContext -> {
+                        return userStore.getSize(storagePath);
+                    });
                     depositTotalSize += fileSize;
 
                 } catch (Exception e) {
@@ -539,8 +528,6 @@ public class Deposit extends Task {
                 }
 
             }
-
-//		depositTotalSize += this.calculateUserUploads(depositTotalSize, context);
 
             // Store the calculated deposit size
             eventSender.send(new ComputedSize(jobID, depositId, depositTotalSize)
@@ -553,26 +540,6 @@ public class Deposit extends Task {
             throw new RuntimeException(e);
         }
 	}
-
-	// TO REMOVE
-	// wpetit: I'm commenting this out as I don't think we want to keep it.
-    // It seems to be necessary when we still had the "Browse" button on the Deposit page
-    // But we removed the button and now is cause the Deposit size to be doubled
-//	private long calculateUserUploads(long depositSize, Context context) {
-//		// Add size of any user uploads
-//        try {
-//            for (String path : fileUploadPaths) {
-//                depositSize += getUserUploadsSize(context.getTempDir(), userID, path);
-//            }
-//        } catch (Exception e) {
-//            String msg = "Deposit failed: could not access user uploads";
-//            logger.error(msg, e);
-//            eventSender.send(new Error(jobID, depositId, msg).withUserId(userID));
-//            throw new RuntimeException(e);
-//        }
-//
-//        return depositSize;
-//	}
 	
 	private HashMap<String, UserStore> setupUserFileStores(StorageClassNameResolver resolver) {
 		userStores = new HashMap<>();
@@ -696,18 +663,6 @@ public class Deposit extends Task {
             logger.debug("Creating chunk checksum task:" + chunkNumber);
             executor.add(ct);
 
-//            chunksHash[i] = Verify.getDigest(chunk);
-//
-//            long chunkSize = chunk.length();
-//            logger.info("Chunk file " + i + ": " + chunkSize + " bytes");
-//            logger.info("Chunk file location: " + chunk.getAbsolutePath());
-//            logger.info("Checksum algorithm: " + tarHashAlgorithm);
-//            logger.info("Checksum: " + chunksHash[i]);
-//
-//            //if (i > (chunkFiles.length / 2)) {
-//            //    throw new Exception("Failed during chunking");
-//            //}
-//            chunksDigest.put(i+1, chunksHash[i]);
         }
 
       executor.execute( result -> {
