+++ datavault-worker/src/main/java/org/datavaultplatform/worker/tasks/Retrieve.java
@@ -18,7 +18,6 @@ import org.datavaultplatform.common.storage.UserStore;
 import org.datavaultplatform.common.storage.Verify;
 import org.datavaultplatform.common.storage.impl.SftpUtils;
 import org.datavaultplatform.common.task.Context;
-import org.datavaultplatform.common.task.Task;
 import org.datavaultplatform.common.task.TaskExecutor;
 import org.datavaultplatform.common.util.StorageClassNameResolver;
 import org.datavaultplatform.common.util.StorageClassUtils;
@@ -36,19 +35,15 @@ import java.time.Clock;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.util.*;
-import java.util.concurrent.TimeUnit;
 
 import org.springframework.retry.RetryCallback;
 
 /**
  * A class that extends Task which is used to handle Retrievals from the vault
  */
-public class Retrieve extends Task {
+public class Retrieve extends BaseTwoSpeedRetryTask {
     public static final String DATA_VAULT_HIDDEN_FILE_NAME = ".datavault";
     private static final String DV_TEMP_DIR_PREFIX = "dvTempDir";
-    private static final int DEFAULT_USERFS_RETRIEVE_ATTEMPTS = 10;
-    private static final int DEFAULT_USERFS_DELAY_SECS_1 = 60;
-    private static final int DEFAULT_USERFS_DELAY_SECS_2 = 60;
     private static final Logger logger = LoggerFactory.getLogger(Retrieve.class);
     private String archiveId = null;
     private String archiveDigestAlgorithm = null;
@@ -64,8 +59,6 @@ public class Retrieve extends Task {
     private String depositCreationDate = null;
     private long archiveSize = 0;
     private EventSender eventSender = null;
-
-    private TwoSpeedRetry userFsTwoSpeedRetry;
 
     /* (non-Javadoc)
      * @see org.datavaultplatform.common.task.Task#performAction(org.datavaultplatform.common.task.Context)
@@ -348,7 +341,11 @@ public class Retrieve extends Task {
         // Check that there's enough free space ...
         final long freespace;
         try {
-            freespace = userFs.getUsableSpace();
+            TwoSpeedRetry.DvRetryTemplate template = userFsTwoSpeedRetry.getRetryTemplate(String.format("calcSizeToFS - %s", retrievePath));
+            freespace = (long)template.execute((RetryCallback<Object, Exception>) retryContext -> {
+                return userFs.getUsableSpace();
+            });
+
         } catch (Exception e) {
             String msg = "Unable to determine free space";
             logger.error(msg, e);
@@ -475,28 +472,7 @@ public class Retrieve extends Task {
         logger.error(msg);
         throw new Exception(msg);
     }
-    protected void setupUserFsTwoSpeedRetry(Map<String, String> properties) {
-        this.userFsTwoSpeedRetry = getUserFsTwoSpeedRetry(properties);
-    }
 
-    protected static TwoSpeedRetry getUserFsTwoSpeedRetry(Map<String, String> properties) {
-
-        int userFsRetrieveMaxAttempts = DEFAULT_USERFS_RETRIEVE_ATTEMPTS;
-        long userFsRetrieveDelayMs1 = TimeUnit.SECONDS.toMillis(DEFAULT_USERFS_DELAY_SECS_1);
-        long userFsRetrieveDelayMs2 = TimeUnit.SECONDS.toMillis(DEFAULT_USERFS_DELAY_SECS_2);
-
-        if (properties.containsKey(PropNames.USER_FS_RETRIEVE_MAX_ATTEMPTS)) {
-            userFsRetrieveMaxAttempts = Integer.parseInt(properties.get(PropNames.USER_FS_RETRIEVE_MAX_ATTEMPTS));
-        }
-        if (properties.containsKey(PropNames.USER_FS_RETRIEVE_DELAY_MS_1)) {
-            userFsRetrieveDelayMs1 = Long.parseLong(properties.get(PropNames.USER_FS_RETRIEVE_DELAY_MS_1));
-        }
-        if (properties.containsKey(PropNames.USER_FS_RETRIEVE_DELAY_MS_2)) {
-            userFsRetrieveDelayMs2 = Long.parseLong(properties.get(PropNames.USER_FS_RETRIEVE_DELAY_MS_2));
-        }
-
-        return new TwoSpeedRetry(userFsRetrieveMaxAttempts, userFsRetrieveDelayMs1, userFsRetrieveDelayMs2);
-    }
     protected void copyFilesToUserFs(Progress progress, File payloadDir, Device userFs, long bagDirSize, String timeStampDirName) throws Exception {
         ProgressTracker tracker = new ProgressTracker(progress, this.jobID, this.depositId, bagDirSize, this.eventSender);
         Thread trackerThread = new Thread(tracker);
